maps/print-maps/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyyaml>=6.0
pydantic>=2.0
redis>=5.0
pillow>=10.0
